requirements.txt
streamlit>=1.29.0
google-generativeai>=0.8.0
langchain>=0.1.0
python-dotenv>=0.19.0
openai>=1.0.0
transformers>=4.36.0
torch>=2.0.0
spacy>=3.7.2
nltk>=3.8.1
fastapi>=0.109.0
uvicorn>=0.27.0
datasets>=2.16.0
evaluate>=0.4.1
scikit-learn>=1.3.2
